Update BarChart stories
