+++ storybook/stories/API/cartesian/Bar.stories.tsx
@@ -1,7 +1,18 @@
 import React from 'react';
 import { Args } from '@storybook/react';
 import { data as dataProp, General as GeneralProps, layout as layoutProp } from '../props/CartesianComponentShared';
-import { ComposedChart, Bar, ResponsiveContainer, Legend, Tooltip, XAxis, YAxis } from '../../../../src';
+import {
+  ComposedChart,
+  Bar,
+  ResponsiveContainer,
+  Legend,
+  Tooltip,
+  XAxis,
+  YAxis,
+  BarChart,
+  CartesianGrid,
+  LabelList,
+} from '../../../../src';
 import { pageData } from '../../data';
 import { getStoryArgsFromArgsTypesObject } from '../props/utils';
 import {
@@ -508,5 +519,100 @@ export const API = {
     animationEasing: 'linear',
     animationBegin: 0,
     animationDuration: 1500,
+  },
+};
+
+const dataWithSmallishValues = [
+  {
+    name: 'Page A',
+    uv: 4000,
+    pv: 2400,
+    amt: 2400,
+  },
+  {
+    name: 'Page B',
+    uv: 3000,
+    pv: 1398,
+    amt: 2210,
+  },
+  {
+    name: 'Page C',
+    uv: 2000,
+    pv: 8,
+    amt: 2290,
+  },
+  {
+    name: 'Page D',
+    uv: 2780,
+    pv: 3908,
+    amt: 2000,
+  },
+  {
+    name: 'Page E',
+    uv: 18,
+    pv: 4800,
+    amt: 2181,
+  },
+  {
+    name: 'Page F',
+    uv: 2390,
+    pv: 3800,
+    amt: 2500,
+  },
+  {
+    name: 'Page G',
+    uv: 3490,
+    pv: 4300,
+    amt: 2100,
+  },
+];
+export const WithMinHeight = {
+  render: (args: Record<string, any>) => {
+    const renderCustomizedLabel = (props: any) => {
+      const { x, y, width, value } = props;
+      const radius = 10;
+
+      return (
+        <g>
+          <circle cx={x + width / 2} cy={y - radius} r={radius} fill="#8884d8" />
+          <text x={x + width / 2} y={y - radius} fill="#fff" textAnchor="middle" dominantBaseline="middle">
+            {value.split(' ')[1]}
+          </text>
+        </g>
+      );
+    };
+
+    return (
+      <ResponsiveContainer width="100%" height="100%">
+        <BarChart
+          width={500}
+          height={300}
+          margin={{
+            top: 20,
+            right: 30,
+            left: 20,
+            bottom: 5,
+          }}
+          data={dataWithSmallishValues}
+        >
+          <CartesianGrid strokeDasharray="3 3" />
+          <XAxis dataKey="name" />
+          <YAxis />
+          <Legend />
+          <Bar {...args}>
+            <LabelList dataKey="name" content={renderCustomizedLabel} />
+          </Bar>
+          <Bar dataKey="uv" fill="#82ca9d" minPointSize={10} />
+          <Tooltip />
+        </BarChart>
+      </ResponsiveContainer>
+    );
+  },
+  args: {
+    ...getStoryArgsFromArgsTypesObject(GeneralProps),
+    ...getStoryArgsFromArgsTypesObject(AnimationProps),
+    dataKey: 'pv',
+    fill: '#8884d8',
+    minPointSize: 5,
   },
 };
