+++ storybook/stories/Examples/ComposedChart/StockPrice.stories.tsx
